+++ atlasdb-client/src/main/java/com/palantir/atlasdb/util/AtlasDbMetrics.java
@@ -34,6 +34,7 @@ import com.palantir.tritium.metrics.MetricRegistries;
 import com.palantir.tritium.metrics.registry.DefaultTaggedMetricRegistry;
 import com.palantir.tritium.metrics.registry.TaggedMetricRegistry;
 import com.palantir.tritium.proxy.Instrumentation;
+import com.palantir.tritium.tracing.TracingInvocationEventHandler;
 
 public final class AtlasDbMetrics {
     private static final Logger log = LoggerFactory.getLogger(AtlasDbMetrics.class);
@@ -88,6 +89,21 @@ public final class AtlasDbMetrics {
         log.warn("Metric Registry was not set, setting to shared default registry name of "
                 + DEFAULT_REGISTRY_NAME);
         return registry;
+    }
+
+    public static <T, U extends T> T instrumentWithTracing(Class<T> serviceInterface, U service) {
+        return instrumentWithTracing(serviceInterface, service, serviceInterface.getName());
+    }
+
+    private static <T, U extends T> T instrumentWithTracing(Class<T> serviceInterface, U service, String name) {
+        return Instrumentation.builder(serviceInterface, service)
+                .withHandler(new MetricsInvocationEventHandler(getMetricRegistry(), name))
+                .withHandler(new TracingInvocationEventHandler(name))
+                .withLogging(
+                        LoggerFactory.getLogger("performance." + name),
+                        LoggingLevel.TRACE,
+                        LoggingInvocationEventHandler.LOG_DURATIONS_GREATER_THAN_1_MICROSECOND)
+                .build();
     }
 
     public static <T, U extends T> T instrument(Class<T> serviceInterface, U service) {