+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/TransactionManagers.java
@@ -293,6 +293,9 @@ public abstract class TransactionManagers {
                 config.initializeAsync(),
                 () -> runtimeConfigSupplier.get().getTimestampCacheSize(),
                 SweepQueueWriter.NO_OP);
+        transactionManager = AtlasDbMetrics.instrumentWithTracing(
+                SerializableTransactionManager.class,
+                transactionManager);
 
         PersistentLockManager persistentLockManager = new PersistentLockManager(
                 persistentLockService,